fp: Refactor code structure
