+++ scripts/flutter/_env.js
@@ -1,0 +1,4 @@
+exports.IF = {
+  useRemote: false,
+  ctx: null
+}
